Rename components to examples https://github.com/infor-design/enterprise-wc/issues/315
