+++ react-ids-wc/src/examples/ids-icon/icons-list.js
