+++ react-ids-wc/src/examples/ids-icon/index.js
