+++ react-ids-wc/src/examples/ids-input/index.js
