+++ react-ids-wc/src/examples/ids-layout-grid/index.js
