+++ react-ids-wc/src/examples/ids-tag/index.js
