+++ react-ids-wc/src/examples/ids-text/index.js
