+++ react-ids-wc/src/examples/index.js
